[#1459] Properly organize CLI commands
**Renamed Files:**
- `validate.rb` → `validate_command.rb`
- `config.rb` → `config_command.rb`
- `initializers.rb` → `initializers_command.rb`
- `migrate.rb` → `migrate_command.rb` (extracted from CLI class)
- `customers.rb` → `customers_command.rb` (extracted from CLI class)
- `domains.rb` → `domains_command.rb` (extracted from CLI class)
- `change_email.rb` → `change_email_command.rb` (extracted from CLI class)
- `move_keys.rb` → `move_keys_command.rb` (extracted from CLI class)
