+++ lib/onetime/cli/initializers_command.rb
