+++ lib/onetime/cli/validate_command.rb
